requirements.txt
# Discord
discord.py>=2.3.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# GitHub
PyGithub>=2.1.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
